@@ -21,7 +21,7 @@ package com.cooltechworks.sample
 
 import android.content.Intent
 import android.os.Bundle
-import android.support.v7.app.AppCompatActivity
+import androidx.appcompat.app.AppCompatActivity
 import android.widget.Button
 import com.cooltechworks.sample.utils.BaseUtils
 import kotlinx.android.synthetic.main.activity_main.*
