@@ -1,10 +1,10 @@
 package com.cooltechworks.sample.utils
 
 
-import android.support.annotation.LayoutRes
-import android.support.annotation.StringRes
-import android.support.annotation.StyleRes
-import android.support.v7.widget.RecyclerView
+import androidx.annotation.LayoutRes
+import androidx.annotation.StringRes
+import androidx.annotation.StyleRes
+import androidx.recyclerview.widget.RecyclerView
 
 class DemoConfiguration {
     @StyleRes
